[sequencing] Change qs BAM tag type from uint8 to float32, conforming to a change made in a dorado commit on 240516.

+++ paulssonlab/src/paulssonlab/sequencing/io.py
@@ -28,7 +28,7 @@ SAM_TAG_REGEX = re.compile(
 DEFAULT_BAM_TAGS = {
     # ONT
     "RG": pa.dictionary(pa.uint8(), pa.string()),
-    "qs": pa.uint8(),
+    "qs": pa.float32(),  # dorado commit on 240516 changed this from int to float
     "ns": pa.int64(),
     "ts": pa.int64(),
     "mx": pa.uint8(),
